+++ src/commands_core/CommandTag.cs
@@ -1,0 +1,8 @@
+namespace MoreCommands.Common;
+
+public enum CommandTag
+{
+    Player,
+    World,
+    Console,
+}