+++ src/common/CommandTag.cs
@@ -1,8 +1,0 @@
-namespace MoreCommands.Common;
-
-public enum CommandTag
-{
-    Player,
-    World,
-    Console,
-}